Move Sync specific code out off Staging
The code on how to sync a file is best located within the Sync command.
The Staging class is only responsible for adding and deleting files from
the manifest and staging directory.

Staging is however still responsible for merging managed files.
