+++ spec/staging_spec.rb
@@ -4,22 +4,6 @@
 require 'staging'
 require 'alces_utils'
 
-module Metalware
-  module Testing
-    class GoodValidation
-      def self.validate(_content)
-        true
-      end
-    end
-
-    class BadValidation
-      def self.validate(_content)
-        false
-      end
-    end
-  end
-end
-
 RSpec.describe Metalware::Staging do
   include AlcesUtils
 
@@ -122,72 +106,6 @@ RSpec.describe Metalware::Staging do
       run_delete_files { |_f| raise 'test error' }
       expect(manifest[:files].length).to eq(files.length)
     end
-  end
-
-  describe '#sync_files' do
-    let :files { ['first', 'second', 'third'].map { |f| "/tmp/#{f}" } }
-    let :validate_file { '/tmp/validate-file' }
-
-    before :each do
-      Metalware::Staging.update(metal_config) do |staging|
-        good_validator = Metalware::Testing::GoodValidation.to_s
-        staging.push_file(validate_file, '', validator: good_validator)
-        files.each { |f| staging.push_file(f, '') }
-      end
-    end
-
-    context 'with valid files (aka no errors expected)' do
-      before :each { Metalware::Staging.update(metal_config, &:sync_files) }
-
-      it 'moves the files into place' do
-        files.each { |f| expect(File.exist?(f)).to eq(true) }
-      end
-
-      it 'leaves the file list empty' do
-        expect(manifest[:files]).to be_empty
-      end
-
-      it 'leaves the staging directory empty' do
-        files = Dir[File.join(file_path.staging_dir, '**/*')].reject do |p|
-          File.directory?(p)
-        end
-        expect(files).to be_empty
-      end
-
-      it 'moves the validated file into place' do
-        expect(File.exist?(validate_file)).to eq(true)
-      end
-    end
-
-    context 'with a validation error' do
-      let :bad_file { '/tmp/bad-validator-file' }
-      let :stderr { StringIO.new }
-
-      before :each do
-        # Allows the error to be printed
-        allow(Metalware::Output).to \
-          (receive(:stderr).and_wrap_original { |_m, *arg| $stderr.puts arg })
-
-        Metalware::Staging.update(metal_config) do |staging|
-          bad_validator = Metalware::Testing::BadValidation.to_s
-          staging.push_file(bad_file, '', validator: bad_validator)
-        end
-        old_stderr = $stderr
-        $stderr = stderr
-        Metalware::Staging.update(metal_config, &:sync_files)
-        $stderr = old_stderr
-      end
-
-      it "isn't removed from the manifest" do
-        expect(manifest[:files].length).to eq(1)
-        expect(manifest[:files].first[0]).to eq(bad_file)
-      end
-
-      it 'issues a validation failure error message' do
-        stderr.rewind
-        expect(stderr.read).to include('ValidationFailure')
-      end
-    end
 
     context 'with a managed file' do
       let :managed_file { '/tmp/managed' }
@@ -200,12 +118,10 @@ RSpec.describe Metalware::Staging do
       before :each do
         Metalware::Staging.update(metal_config) do |staging|
           staging.push_file(managed_file, managed_content, managed: true)
+          staging.delete_file_if do |file|
+            File.write file.sync, file.content
+          end
         end
-        Metalware::Staging.update(&:sync_files)
-      end
-
-      it 'writes to the file if it does not exist' do
-        expect(File.exist?(managed_file)).to eq(true)
       end
 
       it 'writes the managed file content and flags' do
@@ -226,16 +142,15 @@ RSpec.describe Metalware::Staging do
         File.write(managed_file, start_content)
 
         new_content = 'NEW CONTENT'
-        Metalware::Staging.update(metal_config) do |staging|
+        Metalware::Staging.update do |staging|
           staging.push_file(managed_file, new_content, managed: true)
-          staging.sync_files
+          staging.delete_file_if do |file|
+            expect(file.content.first).to eq(file_start)
+            expect(file.content.last).to eq(file_end)
+            expect(file.content).to include(new_content)
+            expect(file.content).not_to include(managed_content)
+          end
         end
-
-        save_content = read_managed_content
-        expect(save_content.first).to eq(file_start)
-        expect(save_content.last).to eq(file_end)
-        expect(save_content).to include(new_content)
-        expect(save_content).not_to include(managed_content)
       end
     end
   end
